requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-multipart>=0.0.9
aiofiles>=23.2.1
pillow>=10.3.0
opencv-python>=4.9.0.80
pytest>=8.0.0
